Added side cta and moved folder
