--- src/pages/blog/side-cta/_data.js
@@ -1,8 +1,0 @@
-import Stocks from './images/stocks.svg'
-
-export const pdf_data = {
-        image: Stocks,
-        title: 'How to trade stocks - the smart way',
-        author: 'Vince Stanzione',
-    }
-
