+++ src/pages/blog/template/_blog-footer/_blog-footer.js
