+++ src/pages/blog/template/_side-cta/_data.js
@@ -1,0 +1,7 @@
+import Stocks from './_images/stocks.svg'
+
+export const pdf_data = {
+    image: Stocks,
+    title: 'How to trade stocks - the smart way',
+    author: 'Vince Stanzione',
+}
